Add basic grid (multiple columns) support
Grids can be defined to create multiple columns at the same time.

The definition is based on three components: keys, a name expression
which uses the key, and a body which determines the value of the cell
based on both the key and the object in the row.

Still needs some treatment of the obtained values and some improvements
in the error treatment.