Homogeneous names for column names and values methods
Every language element that generates a column would have two methods:
getName and getValue. The same happens for those that generate multiple
columns (getNames and getValues).

The reason behind not making this an interface is that not all language
elements need the same attributes to generate names and/or values.