Fix context variable leaks and return values
Variables defined in different column generators could be seen (even in
different datasets!). Now, everything has a local frame as "high" as it
could be placed. Check later if this affects performance.

Also, the ReturnValueParser has been simplified. Now it does not have
execution responsibilities, so that every time the result of a block is
checked it is possible to locally decide if the exception is going to be
thrown or silenced.

In the future, those silenced exceptions should indicate some warning to
the user (to allow they to check wether the error comes from, for
instance, missing data, or from a personal mistake when writing the
column expression).