Introduce from expression in dataset specification
Now it is possible to define a dataset from a list of elements, instead
of iterating all instances of a type and employing a guard.

This also allows to iterate over arbitrary types other than the ones
present in the extracted model, such as integers, strings, or a
structure of them. With this, for instance, group by expression are
possible (not sure if pretty).