Add silent value obtention support
When calculating cell values, it is possible to silent any errors
(introduce blank in cell, continue execution) at module, dataset rule or
column generator level.

The last two ones work through a @silent annotation. The first one can
be set with a method, still needs to be included an option in the run
configuration tab.