Include ColumnGenerator interface abstraction
Now, every element that generates columns in a dataset has to implement
the ColumnGenerator interface, that defines the methods to generate
column names and values.

Each column generator receives the parameters it needs to work through
the initialise method in DatasetRule.

Right now, the getValues() method returns a list of strings that
represent the values. Future changes will make this method to return a
list of objects (versatility when, for instance, post-processing the
values through normalization and such).