Add nested from generator
Now it is possible to nest generators in a nested from expression that
uses a calculated element.

From the dataset param element, a from expression is calculated. Using
the result of that expression, new columns can be defined, using for
that any of the available generators (including nested from into nested
froms).

See example 15 on dummy.lvl for details.